tools/requirements.txt
aiofiles
aiohttp
yarl
Pillow
